Used inheritance to share blurred background functionality
