+++ app/src/main/java/com/codefororlando/streettrees/requesttree/ContactInfoFragment.java
@@ -22,7 +22,7 @@ import com.codefororlando.streettrees.view.PageFragment;
 /**
  * Created by johnli on 9/24/16.
  */
-public class ContactInfoFragment extends PageFragment {
+public class ContactInfoFragment extends BlurredBackgroundFragment {
 
     private Button nextButton;
     private EditText nameField;
@@ -42,7 +42,8 @@ public class ContactInfoFragment extends PageFragment {
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.request_tree_contact_info, container, false);
         bindUi(view);
-        initBlurredBackground(view);
+        initBlurredBackground(view, R.drawable.bg_forrest, 25f, .05f);
+
         return view;
     }
 
@@ -58,13 +59,6 @@ public class ContactInfoFragment extends PageFragment {
                 nextFragment();
             }
         });
-    }
-
-    private void initBlurredBackground(View view) {
-        Bitmap largeIcon = BitmapFactory.decodeResource(getResources(), R.drawable.bg_forrest);
-        Bitmap blurredBackground = BlurBuilder.blur(getActivity(), largeIcon, .05f, 25);
-        Drawable d = new BitmapDrawable(getResources(), blurredBackground);
-        view.setBackground(d);
     }
 
     private void nextFragment() {
