+++ app/src/main/java/com/codefororlando/streettrees/requesttree/RequestIntroFragment.java
@@ -1,9 +1,5 @@
 package com.codefororlando.streettrees.requesttree;
 
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.drawable.BitmapDrawable;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
@@ -12,13 +8,11 @@ import android.view.ViewGroup;
 import android.widget.Button;
 
 import com.codefororlando.streettrees.R;
-import com.codefororlando.streettrees.bitmap.BlurBuilder;
-import com.codefororlando.streettrees.view.PageFragment;
 
 /**
  * Created by johnli on 9/24/16.
  */
-public class RequestIntroFragment extends PageFragment {
+public class RequestIntroFragment extends BlurredBackgroundFragment {
 
     Button nextButton;
 
@@ -34,15 +28,8 @@ public class RequestIntroFragment extends PageFragment {
             }
         });
 
-        initBlurredBackground(view);
+        initBlurredBackground(view, R.drawable.bg_house_center_trees, 25f, .05f);
         return view;
-    }
-
-    private void initBlurredBackground(View view) {
-        Bitmap largeIcon = BitmapFactory.decodeResource(getResources(), R.drawable.bg_house_center_trees);
-        Bitmap blurredBackground = BlurBuilder.blur(getActivity(), largeIcon, .05f, 25);
-        Drawable d = new BitmapDrawable(getResources(), blurredBackground);
-        view.setBackground(d);
     }
 
     @Override
